+++ gt_manage/miniprogram/components/listGoods/listGoods.js
@@ -1,0 +1,29 @@
+// components/listGoods/listGoods.js
+Component({
+  /**
+   * 组件的属性列表
+   */
+  options: {
+    addGlobalClass: true,
+  },
+  properties: {
+    type:{
+      type: String,
+      value: "none"
+    }
+  },
+
+  /**
+   * 组件的初始数据
+   */
+  data: {
+
+  },
+
+  /**
+   * 组件的方法列表
+   */
+  methods: {
+
+  }
+})