+++ gt_manage/miniprogram/components/listOrder/listOrder.js
@@ -1,0 +1,29 @@
+// components/listOrder/listOrder.js
+Component({
+  /**
+   * 组件的属性列表
+   */
+  options: {
+    addGlobalClass: true,
+  },
+  properties: {
+    orderArray:{
+      type: Array,
+      value: [1, 2]
+    }
+  },
+
+  /**
+   * 组件的初始数据
+   */
+  data: {
+
+  },
+
+  /**
+   * 组件的方法列表
+   */
+  methods: {
+
+  }
+})